[FLINK-37730] Moves exception emitter to JobStatusObserver
