+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/observer/deployment/AbstractFlinkDeploymentObserver.java
@@ -74,10 +74,6 @@ public abstract class AbstractFlinkDeploymentObserver
             observeClusterInfo(ctx);
         }
 
-        if (!ReconciliationUtils.isJobInTerminalState(flinkDep.getStatus())) {
-            observeJobManagerExceptions(ctx);
-        }
-
         clearErrorsIfDeploymentIsHealthy(flinkDep);
     }
 
@@ -301,11 +297,4 @@ public abstract class AbstractFlinkDeploymentObserver
      * @param ctx the context with which the operation is executed
      */
     protected abstract void observeFlinkCluster(FlinkResourceContext<FlinkDeployment> ctx);
-
-    /**
-     * Observe the exceptions raised in the job manager and take appropriate action.
-     *
-     * @param ctx the context with which the operation is executed
-     */
-    protected abstract void observeJobManagerExceptions(FlinkResourceContext<FlinkDeployment> ctx);
 }
