+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/observer/deployment/ApplicationObserver.java
@@ -17,9 +17,6 @@
 
 package org.apache.flink.kubernetes.operator.observer.deployment;
 
-import org.apache.flink.annotation.VisibleForTesting;
-import org.apache.flink.api.common.JobID;
-import org.apache.flink.autoscaler.utils.DateTimeUtils;
 import org.apache.flink.kubernetes.operator.api.FlinkDeployment;
 import org.apache.flink.kubernetes.operator.api.status.FlinkDeploymentStatus;
 import org.apache.flink.kubernetes.operator.controller.FlinkResourceContext;
@@ -27,37 +24,11 @@ import org.apache.flink.kubernetes.operator.observer.ClusterHealthObserver;
 import org.apache.flink.kubernetes.operator.observer.JobStatusObserver;
 import org.apache.flink.kubernetes.operator.observer.SnapshotObserver;
 import org.apache.flink.kubernetes.operator.utils.EventRecorder;
-import org.apache.flink.runtime.rest.messages.JobExceptionsInfoWithHistory;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.time.Instant;
-import java.time.ZoneId;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 import static org.apache.flink.kubernetes.operator.config.KubernetesOperatorConfigOptions.OPERATOR_CLUSTER_HEALTH_CHECK_ENABLED;
 
 /** The observer of {@link org.apache.flink.kubernetes.operator.config.Mode#APPLICATION} cluster. */
 public class ApplicationObserver extends AbstractFlinkDeploymentObserver {
-
-    /** The cache entry for the last recorded exception timestamp. */
-    public static final class ExceptionCacheEntry {
-        final String jobId;
-        final long lastTimestamp;
-
-        ExceptionCacheEntry(String jobId, long lastTimestamp) {
-            this.jobId = jobId;
-            this.lastTimestamp = lastTimestamp;
-        }
-    }
-
-    @VisibleForTesting
-    final Map<String, ExceptionCacheEntry> lastRecordedExceptionCache = new ConcurrentHashMap<>();
-
-    private static final Logger LOG = LoggerFactory.getLogger(ApplicationObserver.class);
 
     private final SnapshotObserver<FlinkDeployment, FlinkDeploymentStatus> savepointObserver;
     private final JobStatusObserver<FlinkDeployment> jobStatusObserver;
@@ -82,149 +53,6 @@ public class ApplicationObserver extends AbstractFlinkDeploymentObserver {
                 clusterHealthObserver.observe(ctx);
             }
         }
-    }
-
-    @Override
-    protected void observeJobManagerExceptions(FlinkResourceContext<FlinkDeployment> ctx) {
-        var deployment = ctx.getResource();
-        var operatorConfig = ctx.getOperatorConfig();
-        var jobStatus = deployment.getStatus().getJobStatus();
-        // Ideally should not happen
-        if (jobStatus == null || jobStatus.getJobId() == null) {
-            LOG.warn(
-                    "No jobId found for deployment '{}', skipping exception observation.",
-                    deployment.getMetadata().getName());
-            return;
-        }
-
-        try {
-            JobID jobId = JobID.fromHexString(jobStatus.getJobId());
-            // TODO: Ideally the best way to restrict the number of events is to use the query param
-            // `maxExceptions`
-            //  but the JobExceptionsMessageParameters does not expose the parameters and nor does
-            // it have setters.
-            var history =
-                    ctx.getFlinkService()
-                            .getJobExceptions(
-                                    deployment, jobId, ctx.getDeployConfig(deployment.getSpec()));
-
-            if (history == null || history.getExceptionHistory() == null) {
-                return;
-            }
-
-            var exceptionHistory = history.getExceptionHistory();
-            var exceptions = exceptionHistory.getEntries();
-            if (exceptions.isEmpty()) {
-                return;
-            }
-
-            if (exceptionHistory.isTruncated()) {
-                LOG.warn(
-                        "Job exception history is truncated for jobId '{}'. Some exceptions may be missing.",
-                        jobId);
-            }
-
-            String resourceKey =
-                    deployment.getMetadata().getNamespace()
-                            + "/"
-                            + deployment.getMetadata().getName();
-            String currentJobId = jobStatus.getJobId();
-            Instant lastRecorded = null; // first reconciliation
-
-            ExceptionCacheEntry cacheEntry = lastRecordedExceptionCache.get(resourceKey);
-            if (cacheEntry != null && cacheEntry.jobId.equals(currentJobId)) {
-                lastRecorded = Instant.ofEpochMilli(cacheEntry.lastTimestamp);
-            }
-
-            Instant now = Instant.now();
-            int maxEvents = operatorConfig.getReportedExceptionEventsMaxCount();
-            int maxStackTraceLines = operatorConfig.getReportedExceptionEventsMaxStackTraceLength();
-
-            int count = 0;
-            for (var exception : exceptions) {
-                Instant exceptionTime = Instant.ofEpochMilli(exception.getTimestamp());
-                if (lastRecorded != null && exceptionTime.isBefore(lastRecorded)) {
-                    continue;
-                }
-
-                emitJobManagerExceptionEvent(ctx, exception, exceptionTime, maxStackTraceLines);
-                if (++count >= maxEvents) {
-                    break;
-                }
-            }
-            lastRecordedExceptionCache.put(
-                    resourceKey, new ExceptionCacheEntry(currentJobId, now.toEpochMilli()));
-
-        } catch (Exception e) {
-            LOG.warn(
-                    "Failed to fetch JobManager exception info for deployment '{}'.",
-                    deployment.getMetadata().getName(),
-                    e);
-        }
-    }
-
-    private void emitJobManagerExceptionEvent(
-            FlinkResourceContext<FlinkDeployment> ctx,
-            JobExceptionsInfoWithHistory.RootExceptionInfo exception,
-            Instant exceptionTime,
-            int maxStackTraceLines) {
-
-        String exceptionName = exception.getExceptionName();
-        if (exceptionName == null || exceptionName.isBlank()) {
-            return;
-        }
-
-        Map<String, String> annotations = new HashMap<>();
-        annotations.put(
-                "event-time-readable",
-                DateTimeUtils.readable(exceptionTime, ZoneId.systemDefault()));
-        annotations.put("event-timestamp-millis", String.valueOf(exceptionTime.toEpochMilli()));
-
-        if (exception.getTaskName() != null) {
-            annotations.put("task-name", exception.getTaskName());
-        }
-        if (exception.getEndpoint() != null) {
-            annotations.put("endpoint", exception.getEndpoint());
-        }
-        if (exception.getTaskManagerId() != null) {
-            annotations.put("tm-id", exception.getTaskManagerId());
-        }
-
-        if (exception.getFailureLabels() != null) {
-            exception
-                    .getFailureLabels()
-                    .forEach((k, v) -> annotations.put("failure-label-" + k, v));
-        }
-
-        StringBuilder eventMessage = new StringBuilder(exceptionName);
-        String stacktrace = exception.getStacktrace();
-        if (stacktrace != null && !stacktrace.isBlank()) {
-            String[] lines = stacktrace.split("\n");
-            eventMessage.append("\n\nStacktrace (truncated):\n");
-            for (int i = 0; i < Math.min(maxStackTraceLines, lines.length); i++) {
-                eventMessage.append(lines[i]).append("\n");
-            }
-            if (lines.length > maxStackTraceLines) {
-                eventMessage
-                        .append("... (")
-                        .append(lines.length - maxStackTraceLines)
-                        .append(" more lines)");
-            }
-        }
-
-        String keyMessage =
-                exceptionName.length() > 128 ? exceptionName.substring(0, 128) : exceptionName;
-
-        eventRecorder.triggerEventOnceWithAnnotations(
-                exceptionTime.toEpochMilli(),
-                ctx.getResource(),
-                EventRecorder.Type.Warning,
-                EventRecorder.Reason.JobException,
-                eventMessage.toString().trim(),
-                EventRecorder.Component.JobManagerDeployment,
-                "jobmanager-exception-" + keyMessage.hashCode(),
-                ctx.getKubernetesClient(),
-                annotations);
     }
 
     private class ApplicationJobObserver extends JobStatusObserver<FlinkDeployment> {
