+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/service/FlinkService.java
@@ -21,6 +21,7 @@ import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.client.program.rest.RestClusterClient;
 import org.apache.flink.configuration.Configuration;
+import org.apache.flink.kubernetes.operator.api.AbstractFlinkResource;
 import org.apache.flink.kubernetes.operator.api.FlinkDeployment;
 import org.apache.flink.kubernetes.operator.api.FlinkSessionJob;
 import org.apache.flink.kubernetes.operator.api.spec.FlinkSessionJobSpec;
@@ -129,7 +130,8 @@ public interface FlinkService {
     RestClusterClient<String> getClusterClient(Configuration conf) throws Exception;
 
     JobExceptionsInfoWithHistory getJobExceptions(
-            FlinkDeployment deployment, JobID jobId, Configuration deployConfig) throws Exception;
+            AbstractFlinkResource resource, JobID jobId, Configuration deployConfig)
+            throws Exception;
 
     /** Result of a cancel operation. */
     @AllArgsConstructor
