+++ flink-kubernetes-operator/src/test/java/org/apache/flink/kubernetes/operator/autoscaler/TestingKubernetesAutoscalerUtils.java
@@ -54,6 +54,7 @@ public class TestingKubernetesAutoscalerUtils {
                         },
                         null,
                         new FlinkConfigManager(new Configuration()),
+                        null,
                         null));
     }
 }
