+++ helm/flink-kubernetes-operator/crds/flinkdeployments.flink.apache.org-v1.yml
@@ -1,4 +1,4 @@
-helm/flink-kubernetes-operator/crds/flinkdeployments.flink.apache.org-v1.yml# Generated by Fabric8 CRDGenerator, manual edits might get overwritten!
+# Generated by Fabric8 CRDGenerator, manual edits might get overwritten!
 apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
